all changes

+++ divar/ads_app/models.py
@@ -1,5 +1,6 @@
 from django.db import models
 from user_app.models import UserAccount
+from django.core.validators import MinValueValidator
 
 class Category(models.Model):
     name = models.CharField(max_length=100)
@@ -10,7 +11,7 @@ class Category(models.Model):
 class Ad(models.Model):
     title = models.CharField(max_length=100)
     description = models.TextField()
-    price = models.IntegerField()
+    price = models.PositiveBigIntegerField()
     location = models.CharField(max_length=255)
     category = models.ForeignKey(Category , on_delete=models.CASCADE)
     user = models.ForeignKey(UserAccount, on_delete=models.CASCADE)
